replace jsp to json, add new service methods

+++ src/main/com/intexsoft/test_web_mvc/service/impl/SubscriberServiceImpl.java
@@ -27,7 +27,7 @@ public class SubscriberServiceImpl implements SubscriberService {
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         subscriberRepository.deleteById(id);
     }
